// src/libs/ifield.interface.ts
export interface IType { 
  defaultValue: any; 
  name: string; 
  //nestedType?: IType[] | {[key:string]:IType}; 
  enums?: string[]; // ?? 
  isEnum?: boolean; 
  isArray?: boolean; 
  isScalar?: boolean; 
  isObject?: boolean; 
} 

export interface IOption { 
  value:any; 
  label:string; 
}

export interface IField {
  name: string; 
  type: IType; 

  label: string; 
  ref?: string; 
  options?: any; 
  abbrev: boolean; 
  // required?: boolean; 
  // unique?: boolean; 
  // regex: string ?? 
  // format: string ?? 
  // validators: any[];
  // abbrev: 
  
}

export interface IModel { 
  accessor:string, 
  label:string[], 
  description:string[], 
  fields:IField[], 
}; 

export interface IEntry { 
  _id: string; 
  [key:string]: any; 
} 